chore : 동작 과정 수정
