+++ backend/src/main/java/com/hjjang/backend/domain/image/dto/ImageResponse.java
@@ -1,0 +1,2 @@
+package com.hjjang.backend.domain.image.dto;public class ImageResponse {
+}
