+++ backend/src/main/java/com/hjjang/backend/domain/image/exception/ImageNotFoundException.java
@@ -1,0 +1,2 @@
+package com.hjjang.backend.domain.image.exception;public class ImageNotFoundException {
+}
